refactor(Line,Point,Ladder): 기존 객체 삭제 후 새로운 클래스 이름 수정
