+++ src/test/java/nextstep/ladder/LadderTest.java
@@ -13,14 +13,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class LadderTest {
-    @Test
-    @DisplayName("입력받은 참가자 이름(width - 1)과 사다리 높이(height) 만큼의 사다리가 생성되는지 확인한다.")
-    void createLadder() {
-        Ladder ladder = new Ladder(4, 5);
-
-        assertThat(ladder.getWidth()).isEqualTo(3);
-        assertThat(ladder.getHeight()).isEqualTo(5);
-    }
 
     @Test
     @DisplayName("사다리 높이가 1보다 작을 경우 Exception 처리를 한다.")
@@ -29,11 +21,11 @@ public class LadderTest {
 
         assertThatIllegalArgumentException().isThrownBy(() -> {
             new Ladder(4, height);
-    }).withMessage(Ladder.LADDER_HEIGHT_ERROR);
-}
+        }).withMessage(Ladder.LADDER_HEIGHT_ERROR);
+    }
 
     @Test
-    @DisplayName("사다리의 Point(참여자 수 - 1)는 2명 이상이어야 한다.")
+    @DisplayName("사다리의 참여자 수(Point)는 2 이상이어야 한다.")
     void assertLadderWidth() {
         int participantCount = 1;
 
@@ -48,8 +40,7 @@ public class LadderTest {
      * |     |-----|
      */
     @Test
-    @DisplayName("출발점으로부터 사다리를 타고 알맞은 Position에 도착해야 한다.")
-    void moveTest1() {
+    void move() {
         List<Line> lines = new ArrayList<>();
 
         lines.add(createHasLineLeft());
@@ -64,25 +55,31 @@ public class LadderTest {
     }
 
     /**
-     * |-----|     |
+     * |     |-----|
      */
-    private static Line createHasLineLeft() {
+    private static Line createHasLineRight() {
         List<Point> points = new ArrayList<>();
-
-        points.add(new Point(0, true));
-        points.add(new Point(1, false));
+        Point point = Point.first(false);
+        points.add(point);
+        point = point.next(true);
+        points.add(point);
+        point = point.last();
+        points.add(point);
 
         return new Line(points);
     }
 
     /**
-     * |     |-----|
+     * |-----|     |
      */
-    private static Line createHasLineRight() {
+    private static Line createHasLineLeft() {
         List<Point> points = new ArrayList<>();
-
-        points.add(new Point(0, false));
-        points.add(new Point(1, true));
+        Point point = Point.first(true);
+        points.add(point);
+        point = point.next();
+        points.add(point);
+        point = point.last();
+        points.add(point);
 
         return new Line(points);
     }
