+++ src/test/java/nextstep/ladder/PointTest.java
@@ -1,6 +1,6 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.Point2;
+import nextstep.ladder.domain.Point;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,15 +8,15 @@ import static java.lang.Boolean.FALSE;
 import static java.lang.Boolean.TRUE;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class Point2Test {
+public class PointTest {
     /**
      * startPosition: 0
      * |-----|
      */
     @Test
     public void first() {
-        assertThat(Point2.first(TRUE).move()).isEqualTo(1);
-        assertThat(Point2.first(FALSE).move()).isEqualTo(0);
+        assertThat(Point.first(TRUE).move()).isEqualTo(1);
+        assertThat(Point.first(FALSE).move()).isEqualTo(0);
     }
 
     /**
@@ -25,7 +25,7 @@ public class Point2Test {
      */
     @Test
     public void next_stay() {
-        Point2 second = Point2.first(FALSE).next(FALSE);
+        Point second = Point.first(FALSE).next(FALSE);
         assertThat(second.move()).isEqualTo(1);
     }
 
@@ -35,7 +35,7 @@ public class Point2Test {
      */
     @Test
     public void next_left() {
-        Point2 second = Point2.first(TRUE).next(FALSE);
+        Point second = Point.first(TRUE).next(FALSE);
         assertThat(second.move()).isEqualTo(0);
     }
 
@@ -45,7 +45,7 @@ public class Point2Test {
      */
     @Test
     public void next_right() {
-        Point2 second = Point2.first(FALSE).next(TRUE);
+        Point second = Point.first(FALSE).next(TRUE);
         assertThat(second.move()).isEqualTo(2);
     }
 
@@ -56,7 +56,7 @@ public class Point2Test {
     @Test
     @DisplayName("첫번째 Point가 Right true이면, 다음 Point는 무조건 Right false 여야 한다.(가로라인이 겹치는 경우 어느 방향으로 이동할 지 결정할 수 없다.)")
     public void next() {
-        Point2 second = Point2.first(TRUE).next();
+        Point second = Point.first(TRUE).next();
         assertThat(second.move()).isEqualTo(0);
     }
 }
